feat: Organised code into seperate files
Game runs as did before.
Organised into more readable files.
