+++ settings.py
@@ -1,0 +1,13 @@
+# Genral Game settings
+GAME_WIDTH = 700
+GAME_HEIGHT = 700
+SPEED = 100
+SPACE_SIZE = 50
+BODY_PARTS = 3
+SNAKE_COLOR = "#00FF00"
+FOOD_COLOR = "#FF0000"
+BACKGROUND_COLOR = "#000000"
+
+# Game variables
+score = 0
+direction = 'down'
